Bugfix.
A bug was detected in `AssignmentCompressor` where, if the number of
compressed items was an integer multiple of `window`, an additional
assignment was read out when decompressing.

This bug was borne of an oversight in teardown logic. When the user's
`with` statement is exited, the `AssignmentCompressor` calls its
`__exit__()` method, allowing for safe tears-down. Previously, when this
method was called, `_compress(force=True)` was called, indicating that
the remaining items in the cache should be forcibly compressed *without*
the max cache window being hit, and written to file with a chunk
delimiter (i.e. `AssignmentCompressor.CHUNK_SEPARATOR`). However, this
teardown logic did not check whether the cache had items awaiting
compression; if the cache was empty and `force=True` was passed anyway,
an empty bytestring trailed the chunk delimiter.

Because the decompression algorithm looks for the chunk delimiter when
decompressing (to decide when a chunk of plans has ended, and that chunk
of plans should be decompressed and `yield`ed), it treated the trailing
empty bytestring as an assignment. As a result, the empty bytestring was
translated into a dictionary with one key and no value and `yield`ed,
resulting in the user reading out one more assignment than they read in.

Now, teardown logic verifies whether the cache has items awaiting
compression before calling `_compress(force=True)`; if not, we do not
forcibly compress, and cleanly stop compressing.